Move generate message id to file
